requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
requests
squareup
boto3
python-multipart
pydantic[email]